WIP: feat: polyphony with fundsp
Changed audio engine to fundsp, which supports better tooling for
syntesis, sampling and out of the box polyphony.

I suspect some type of wait in the main loop causes ALSA to underrun (as there's
not enough audio to play). TODO: write a tiny example and try to build
up from there by adding the TUI